Game Automation/requirements.txt
cv2
cvzone.HandTrackingModule 
pyautogui
time
threading
playsound 